Refactor CodingTransformer into modular files
Split CodingTransformer.swift into three files: CodingTransformer.swift (protocols and extensions), CodingTransformerComposition.swift (composed transformers), and BuiltInTransformers.swift (built-in transformers). This improves code organization and maintainability by separating protocol definitions, composition logic, and built-in implementations.
